[receivers/promtail] update loki, address review comments
