+++ receiver/promtailreceiver/promtail_test.go
@@ -1,74 +1,87 @@
 package promtailreceiver
 
 import (
-	"path/filepath"
 	"testing"
 	"time"
 
-	"github.com/grafana/loki/clients/pkg/promtail/targets/file"
-
+	"github.com/grafana/loki/clients/pkg/promtail/api"
 	"github.com/grafana/loki/clients/pkg/promtail/scrapeconfig"
+	"github.com/grafana/loki/clients/pkg/promtail/targets/file"
+	"github.com/grafana/loki/pkg/logproto"
+	"github.com/open-telemetry/opentelemetry-collector-contrib/pkg/stanza/entry"
+	"github.com/open-telemetry/opentelemetry-collector-contrib/pkg/stanza/testutil"
 	"github.com/prometheus/common/model"
-	"github.com/prometheus/prometheus/discovery"
-
-	"github.com/grafana/loki/clients/pkg/promtail/positions"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/pkg/stanza/adapter"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/pkg/stanza/operator"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/pkg/stanza/operator/input/promtail"
-	"github.com/stretchr/testify/assert"
+	"github.com/prometheus/prometheus/discovery/targetgroup"
 	"github.com/stretchr/testify/require"
-	"go.opentelemetry.io/collector/config"
-	"go.opentelemetry.io/collector/confmap/confmaptest"
 )
 
-func TestLoadConfig(t *testing.T) {
-	cm, err := confmaptest.LoadConf(filepath.Join("testdata", "config.yaml"))
-	require.NoError(t, err)
-
-	factory := NewFactory()
-	cfg := factory.CreateDefaultConfig()
-
-	sub, err := cm.Sub(config.NewComponentID("promtail").String())
-	require.NoError(t, err)
-	require.NoError(t, config.UnmarshalReceiver(sub, cfg))
-
-	assert.NoError(t, cfg.Validate())
-	bla := testdataConfigYaml()
-	assert.Equal(t, bla, cfg)
-}
-
-func testdataConfigYaml() *PromtailConfig {
-	return &PromtailConfig{
-		BaseConfig: adapter.BaseConfig{
-			ReceiverSettings: config.NewReceiverSettings(config.NewComponentID(typeStr)),
-			Operators:        []operator.Config{},
-		},
-		InputConfig: func() promtail.Config {
-			c := promtail.NewConfig()
-			c.PositionsConfig = positions.Config{PositionsFile: "/tmp/positions.yaml"}
-			c.ScrapeConfig = []scrapeconfig.Config{
-				{
-					JobName: "system",
-					ServiceDiscoveryConfig: scrapeconfig.ServiceDiscoveryConfig{
-						StaticConfigs: discovery.StaticConfig{
-							{
-								Labels: model.LabelSet{
-									"job":      "varlogs",
-									"__path__": "testdata/simple.log",
-								},
+func TestPromtailInput_parsePromtailEntry(t *testing.T) {
+	basicConfig := func() *Config {
+		cfg := NewConfigWithID("testfile")
+		cfg.ScrapeConfig = []scrapeconfig.Config{
+			{
+				JobName: "testjob",
+				ServiceDiscoveryConfig: scrapeconfig.ServiceDiscoveryConfig{
+					StaticConfigs: []*targetgroup.Group{
+						{
+							Labels: model.LabelSet{
+								"job":    "varlogs",
+								"__path": "/var/log/example.log",
 							},
 						},
 					},
 				},
-			}
+			},
+		}
+		cfg.TargetConfig = file.Config{
+			SyncPeriod: 10 * time.Second,
+		}
 
-			syncPeriod, _ := time.ParseDuration("10s")
+		return cfg
+	}
 
-			c.TargetConfig = file.Config{
-				SyncPeriod: syncPeriod,
-			}
+	cfg := basicConfig()
+	op, err := cfg.Build(testutil.Logger(t))
+	require.NoError(t, err)
 
-			return *c
-		}(),
+	promtailInput := op.(*PromtailInput)
+
+	cases := []struct {
+		name        string
+		inputEntry  api.Entry
+		outputEntry entry.Entry
+	}{
+		{
+			name: "Success",
+			inputEntry: api.Entry{
+				Labels: model.LabelSet{
+					"filename": "/var/log/example.log",
+					"job":      "varlogs",
+				},
+				Entry: logproto.Entry{
+					Timestamp: time.Now(),
+					Line:      "test message",
+				},
+			},
+			outputEntry: entry.Entry{
+				Body: "test message",
+				Attributes: map[string]interface{}{
+					"log.file.name": "example.log",
+					"log.file.path": "/var/log/example.log",
+					"job":           "varlogs",
+				},
+			},
+		},
+	}
+
+	for _, tc := range cases {
+		t.Run(
+			tc.name, func(t *testing.T) {
+				outputEntry, err := promtailInput.parsePromtailEntry(tc.inputEntry)
+				require.NoError(t, err)
+				require.Equal(t, tc.outputEntry.Body, outputEntry.Body)
+				require.Equal(t, tc.outputEntry.Attributes, outputEntry.Attributes)
+			},
+		)
 	}
 }
