+++ receiver/promtailreceiver/testdata/config.yaml
